+++ Example/Program.cs
@@ -19,17 +19,28 @@ class Program
         int pointAmount = 0;
 
         Parakolay apiClient = new Parakolay(baseUrl, apiKey, apiSecret, merchantNumber, conversationId, clientIpAddress);
-        Init3dsResponseModel result =new Init3dsResponseModel();
 
-        if (args.Length == 0 || args[0] != "return")
+        if (args.Length == 0)
         {
-            result = await apiClient.Init3DS("CARD_NUMBER", "CARDHOLDER_NAME", "EXPIRE_MONTH (MM)", "EXPIRE_YEAR (YY)", "CVV", amount, pointAmount, "YOUR_CALLBACK_URL");
-            Console.WriteLine(result);
+            Init3dsResponseModel result = await apiClient.Init3DS("CARD_NUMBER", "CARDHOLDER_NAME", "EXPIRE_MONTH (MM)", "EXPIRE_YEAR (YY)", "CVV", amount, pointAmount, "YOUR_CALLBACK_URL");
+            Console.WriteLine(result.cardToken);
+            Console.WriteLine(result.threeDSessionID);
+            Console.WriteLine(result.htmlContent);
         }
-        else
+        else if (args[0] == "complete")
         {
-            var complete = apiClient.Complete3DS(result);
+            ProvisionResult complete = await apiClient.Complete3DS("3DS_SESSION_ID", amount, "CARDHOLDER_NAME", "CARD_TOKEN");
             Console.WriteLine(complete);
+        }
+        else if (args[0] == "reverse")
+        {
+            ReverseResult reverse = await apiClient.Reverse("ORDER_ID");
+            Console.WriteLine(reverse);
+        }
+        else if (args[0] == "return")
+        {
+            ReturnResult ret = await apiClient.Return(amount, "ORDER_ID");
+            Console.WriteLine(ret);
         }
     }
 }